backend/requirements.txt
# Core dependencies
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0
email-validator==2.0.0

# Database
sqlalchemy==2.0.20
alembic==1.11.3
psycopg==3.1.10
# psycopg-binary==3.1.10  # Removed due to compatibility issues

# Authentication
python-jose==3.3.0
cryptography==41.0.3
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Workers
celery==5.3.1
redis==4.6.0

# Testing
pytest==7.4.0
httpx==0.24.1
pytest-cov==4.1.0

# Utils
python-dotenv==1.0.0
tenacity==8.2.3
pydantic-settings==2.0.3
